requirements.txt
pandas
sqlalchemy
psycopg2-binary
google-cloud-bigquery
python-dotenv
